fix warning

+++ utils.h
@@ -13,7 +13,7 @@ inline std::string generate_random_string(size_t length = 0)
     static const std::string allowed_chars{ "123456789BCDFGHJKLMNPQRSTVWXZbcdfghjklmnpqrstvwxz" };
 
     static thread_local std::default_random_engine randomEngine(std::random_device{}());
-    static thread_local std::uniform_int_distribution<int> randomDistribution(0, allowed_chars.size() - 1);
+    static thread_local std::uniform_int_distribution<size_t> randomDistribution(0, allowed_chars.size() - 1);
 
     std::string id(length ? length : 32, '\0');
